[non-iterable-map] A map adapter that does not allow the usage of begin(), end() from the map to prevent iteration.
Often times in compiler data structures, we use unordered maps to map
pointers to other state in the program. Iterating over such unordered maps
causes the compiler to potentially become non-deterministic. The intention
of this adapter is to provide the compiler writer with a way use the type
system to eliminate this bug by banning iteration over an underlying map
type.

NOTE: When asserts are enabled, we still provide a set of debugBegin(),
debugEnd() entry points that allow for iteration over the underlying
map. The reason for this is to still allow for dump() methods to be
implemented on top of the map.

NOTE: This does violate the Liskov Substitution principle. *BUT* The worst
that could happen is that we miss a case if c++ calls MapTy::begin() instead
of erroring. This is better than the status quo when /all/ uses of the
underlying map can lead to non-determinism.