[update-checkout] Improve compatibility with Python 3
Switch the ambigous relative imports to Python 3 relative imports
(supported in Python 2.7). Only perform the renaming of the main script
in Python 2, or Python 3 will fall into a recursive loop (if the
workaround is removed, Python 2 falls into a similar recursive loop).
Finally add a couple of decode() calls for the output/error of the shell
utilities to be printed as text in Python 3.