[ParseableInterface] Hack: prefer module names over type names
...in the parts of a .swiftinterface file that are pretty-printed.

The effect of this is to dodge a longstanding issue (SR-898) where if
you have a module and type with the same name, the compiler always
prefers the type. This is a problem for swiftinterface files, which
are pretty-printed to use qualified types everywhere (to minimize the
chance of collisions when a library changes). We know that one of
these types exists in the wild, too: XCTest, the base class of
XCTestCase.

This does not fix SR-898; it only works around it in certain limited
cases by making the /opposite/ assumption: preferring a module name
instead of a type name. We also considered having some special syntax
that indicates that the first component of a qualified name is a
module, but felt that that might actually be a feature of the language
some day, and wouldn't want to end up with two syntaxes for the same
thing (or worse, have subtly different semantics for the one used in
swiftinterfaces, or have it conflict with something /else/).