[AST] NFC: Generalize no-escape tracking
This patch generalizes the existing "no-escape" semantic analysis and
makes future features possible. In my case, I have a research branch
with a couple new types that need no-escape behavior. As for Swift in
general, this opens the door to things like "no-escape array of
no-escape functions" or "no-escape tuple of no-escape functions".