[cast-opt] Eliminate always true if statement.
I think that this is a result of previous refactorings. This code has a long
time assert in it that Src (since 2016) is an address and for a long time the bridged type has
always been an object (since 2015). So, thus we know that the following is always true.

  SILBridgedTy != Src->getType()

I also removed a level of indentation. NFC.

The 2015 commit hash is: 2dd38eee0ede4a51f4e7d1bf9272ae91b3a8a1ff (the original commit).
The 2016 commit hash is: 506ab9809f0187c5242d57cadd80e9c07286329e