Changes needed to build Swift toolchain with --test option in release mode.
Invoking the toolchain and the tests using the following default command:-
	./swift/utils/build-toolchain mybundle --test 
This creates a release build by default using the defaults in 'utils/build-presets.ini' file , where source files are compiled without any debug information.
So in that case the binaries(test binaries 'a.out' and others) are not really helpful in debugging with lldb debugger which needs debugging information. Hence the lldb suite tests which involve debugging with lldb debugger, fail.