Frontend: sketch out a SwiftRuntime infrastructure
Due to the ABI stability, it will become increasingly useful to be able
to query the runtime that we are targeting to take advantage of newer
functionality if possible.  Sketch out a framework for this inspired by
the ObjCRuntime system in clang.  Although there is no reason to assume
that there will be variations of the Kind sort, leave the infrastructure
generic enough to support that.

This addresses some comments on the changes to support lazy registration
of foreign import types.