[build-script] Introduce BuildScriptImplBuilder
BuildScriptImplBuilder is a ProductBuilder that invokes
build-script-impl to work. The implementation is moved code from the
main script BuildScriptInvocation type into the new builder.

This will allow us to replace each of the product builders one by one
and check them before commiting to the new system.

There's a little reorganization in the execute method of
BuildScriptInvocation to make it less tied to the whims of
build-script-impl.