[build-script] Transform IndexStoreDB and SourcekitLSP to use ProductBuilders.
Create a new helper for IndexStoreDB and SourcekitLSP called
BuildScriptHelperBuilder. Move most of the code that existed in a
function in indexstoredb.py into that helper object.

Create small builders for IndexStoreDB and SourcekitLSP which use this
helper object as their base.

Modify the main build-script to invoke the new builder instead of the
specific do_build/do_test. This way, all products are using the same
system to be build, but some use build-script-impl, and others use
build-script-helper.py.