[TF-86] using expectEqual to compare tensors and expected values.
It's now possible to use expectEqual instead of expectTrue.
expectEqual would give better error messages in case the comparison
fails.

As well, expectEqual expects the first parameters to be the expected
value so I changed some of the statements to match this assumption.
When printing failure messages getting right expected and actual values
avoids confusions.