IRGen: correct field metadata section name on COFF
COFF has a section name limit which was being exceeded.  Correct the
section name to what was expected (the tests are lenient and would
accept the ELF name as well).  Furthermore, ensure that the section is
emitted into the B group so that it can be collated and markers
inserted.