Update Python Interop ref-counting names and add leak testing. (#23316)
The old names, `borrowing:` and `owning:` mean opposite things from
opposite perspectives. `owning:` may mean that the PyReference owns the
ref and needs to increment it itself, or from a caller perspective it
may mean that it needs to increment it.

To clarify this, we (Dan and I) discussed using `adopting` to signify
that the PyReference adopts a +1 ref count from the calling context.

Likewise, `retaining:` will make no overall change but will retain the
reference for the lifespan of PyReference object.