Avoided double typechecking when possible.
Optimized for the common case where a single expression in a function
body is an implicit return and attempt to solve the constraint system
featuring that assumption.  If that fails, attempt to solve a second
constraint system where no conversion type is provided.  If that
succeeds and the expression's type is uninhabited, then remove the
return and proceed.  If that fails, show the diagnostics from the
failure to solve the first constraint system.