[Property delegates] Implement support for init(initialValue:).
When a property delegate type includes an initializer
init(initialValue:), a property with that delegate can be initialized
directly (with =), and the backing storage will be provided that value
via the .init(initialValue:) initializer. For example, given a trivial
wrapper:

  @propertyDelegate
  struct Wrapper<T> {
    var value: T
    init(initialValue: T) { self.value = initialValue }
  }

  var x by Wrapper = 17

The type checker will infer the type 'Int' for 'x' and, therefore,
'Wrapper<Int>` for the backing property.