[Property delegates] Associate initializer with the backing storage property.
When a property that has a property delegate has an initializer, associate
that initializer with the backing storage property and make it “subsume”
the initializer on the original declaration.

While here, ensure that we emit the accessors for the property that has a
property delegate, and tie it together with a -dump-ast test case to
ensure that we’re getting sensible ASTs out of this.

Examples involving uses of property delegates for instance properties of
types now work end-to-end.