test: improve test coverage rate for Windows
Strip the use of `env` on Windows which will swallow the error code if
the GnuWin32 version (which is recommended/used by LLVM) is used.  Since
SDKROOT should not be set on Windows anyways, this should not matter
much in practice but will help improve the test coverage on Windows.

Additionally, remove the path lowercasing on Windows as the path is kept
in the proper case, even on the nightlies.  This should help improve the
number of Driver tests that pass.

With this change, I now can get all the Driver tests to pass locally.